@@ -12,8 +12,8 @@
     "test": "node test"
   },
   "dependencies": {
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "react-jenny": "^1.1.0"
   },
   "devDependencies": {
